Add module to hold tag-attribute functions and updated contextmenu-layer to use them
